Fix bug with selecting ingredients
In the item selection dialog, it was not possible to properly switch to
an incredient using the recipe view when the item filter filtered
the item list to a single item. The fact that the item list had only
one item overwrote the selected item before the selection through
the click on the ingredient could be applied.